requirements.txt
langchain
fastapi
openai
python-dotenv
uvicorn[standard]
gunicorn
python-multipart
tiktoken
pypdf
pinecone-client